requirements.txt
aioconsole==0.3.2
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
chardet==4.0.0
discord.py==2.0.0a3508+g40650147
idna==3.2
multidict==5.1.0
typing-extensions==3.10.0.0
tzdata==2021.1
yarl==1.6.3
